Fix warning.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5980 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CMS3DMeshFileLoader.cpp
@@ -192,7 +192,7 @@ bool CMS3DMeshFileLoader::load(io::IReadFile* file)
 	// read whole file
 
 	u8* buffer = new u8[fileSize];
-	s32 read = file->read(buffer, fileSize);
+	size_t read = file->read(buffer, fileSize);
 	if (read != fileSize)
 	{
 		delete [] buffer;
